+++ User/Boards/DevC/bsp_pwm.c
@@ -5,13 +5,15 @@
 
 #include "tim.h"
 
+#include "user_math.h"
+
 /* Private define ------------------------------------------------------------*/
 /* Private macro -------------------------------------------------------------*/
 /* Private typedef -----------------------------------------------------------*/
 /* Private variables ---------------------------------------------------------*/
 /* Private function  ---------------------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
-int BSP_PWM_Start(BSP_PWM_Channel_t ch) {
+int8_t BSP_PWM_Start(BSP_PWM_Channel_t ch) {
 	switch(ch) {
 		case BSP_PWM_IMU_HEAT: 
 			HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_2); 
@@ -20,7 +22,7 @@ int BSP_PWM_Start(BSP_PWM_Channel_t ch) {
 	return 0;
 }
 
-int BSP_PWM_Set(BSP_PWM_Channel_t ch, float duty_cycle) {
+int8_t BSP_PWM_Set(BSP_PWM_Channel_t ch, float32_t duty_cycle) {
 	if (duty_cycle > 1.f)
 		return -1;
 	
@@ -37,7 +39,7 @@ int BSP_PWM_Set(BSP_PWM_Channel_t ch, float duty_cycle) {
 	return 0;
 }
 
-int BSP_PWM_Stop(BSP_PWM_Channel_t ch) {
+int8_t BSP_PWM_Stop(BSP_PWM_Channel_t ch) {
 	switch(ch) {
 		case BSP_PWM_IMU_HEAT:
 			HAL_TIM_PWM_Stop(&htim5, TIM_CHANNEL_2);
