+++ User/Boards/bsp_gpio.h
@@ -8,7 +8,7 @@
 /* Exported macro ------------------------------------------------------------*/
 /* Exported types ------------------------------------------------------------*/
 /* Exported functions prototypes ---------------------------------------------*/
-int BSP_GPIO_RegisterCallback(uint16_t pin, void (*callback)(void));
+int8_t BSP_GPIO_RegisterCallback(uint16_t pin, void (*callback)(void));
 
-int BSP_GPIO_EnableIRQ(uint16_t pin);
-int BSP_GPIO_DisableIRQ(uint16_t pin);
+int8_t BSP_GPIO_EnableIRQ(uint16_t pin);
+int8_t BSP_GPIO_DisableIRQ(uint16_t pin);
