+++ User/Boards/bsp_laser.h
@@ -2,12 +2,14 @@
 
 
 /* Includes ------------------------------------------------------------------*/
+#include "user_math.h"
+
 /* Exported constants --------------------------------------------------------*/
 /* Exported macro ------------------------------------------------------------*/
 /* Exported types ------------------------------------------------------------*/
 /* Exported functions prototypes ---------------------------------------------*/
 
 /* duty_cycle大于零时，A板为全开，C板为pwm调光*/
-int BSP_Laser_Start(void);
-int BSP_Laser_Set(float duty_cycle);
-int BSP_Laser_Stop(void);
+int8_t BSP_Laser_Start(void);
+int8_t BSP_Laser_Set(float32_t duty_cycle);
+int8_t BSP_Laser_Stop(void);
